+++ ContextualPopupDecorator/ContextualPopupDecorator.js
@@ -77,7 +77,7 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 	const {noArrow, noSkin, openProp} = config;
 
 	return class extends React.Component {
-		static displayName = 'ContextualPopupDecorator'
+		static displayName = 'ContextualPopupDecorator';
 
 		static propTypes = /** @lends sandstone/ContextualPopupDecorator.ContextualPopupDecorator.prototype */ {
 			/**
@@ -243,7 +243,7 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 			 * @public
 			 */
 			spotlightRestrict: PropTypes.oneOf(['none', 'self-first', 'self-only'])
-		}
+		};
 
 		static defaultProps = {
 			'data-webos-voice-exclusive': true,
@@ -253,7 +253,7 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 			open: false,
 			scrimType: 'none',
 			spotlightRestrict: 'self-first'
-		}
+		};
 
 		constructor (props) {
 			super(props);
@@ -266,6 +266,7 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 
 			this.overflow = {};
 			this.adjustedDirection = this.props.direction;
+			this.id = this.generateId();
 
 			this.MARGIN = ri.scale(noArrow ? 0 : 12);
 			this.ARROW_WIDTH = noArrow ? 0 : ri.scale(60); // svg arrow width. used for arrow positioning
@@ -331,6 +332,10 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 			Spotlight.remove(this.state.containerId);
 		}
 
+		generateId = () => {
+			return Math.random().toString(36).substr(2, 8);
+		};
+
 		getContainerNodeWidth () {
 			return this.containerNode && this.containerNode.getBoundingClientRect().width || 0;
 		}
@@ -359,7 +364,7 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 			}
 
 			return {anchor, direction};
-		}
+		};
 
 		getContainerPosition (containerNode, clientNode) {
 			const position = this.centerContainerPosition(containerNode, clientNode);
@@ -571,17 +576,17 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 					});
 				}
 			}
-		}
+		};
 
 		getContainerNode = (node) => {
 			this.containerNode = node;
-		}
+		};
 
 		getClientNode = (node) => {
 			this.clientNode = ReactDOM.findDOMNode(node); // eslint-disable-line react/no-find-dom-node
-		}
+		};
 
-		handle = handle.bind(this)
+		handle = handle.bind(this);
 
 		handleKeyUp = this.handle(
 			forProp('open', true),
@@ -589,7 +594,7 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 			() => Spotlight.getCurrent() === this.state.activator,
 			stop,
 			forward('onClose')
-		)
+		);
 
 		handleOpen = (ev) => {
 			forward('onOpen', ev, this.props);
@@ -600,14 +605,14 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 				activator: current
 			});
 			this.spotPopupContent();
-		}
+		};
 
 		handleClose = () => {
 			this.updateLeaveFor(null);
 			this.setState({
 				activator: null
 			});
-		}
+		};
 
 		handleDirectionalKey (ev) {
 			// prevent default page scrolling
@@ -640,7 +645,7 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 					this.spotPopupContent();
 				}
 			}
-		}
+		};
 
 		// handle key event from contextual popup and closes the popup
 		handleContainerKeyDown = (ev) => {
@@ -655,7 +660,7 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 			if (Spotlight.move(direction) && !this.containerNode.contains(Spotlight.getCurrent())) {
 				forward('onClose', ev, this.props);
 			}
-		}
+		};
 
 		spotActivator = (activator) => {
 			if (!Spotlight.getPointerMode() && activator && activator === Spotlight.getCurrent()) {
@@ -664,7 +669,7 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 			if (!Spotlight.focus(activator)) {
 				Spotlight.focus();
 			}
-		}
+		};
 
 		spotPopupContent = () => {
 			const {spotlightRestrict} = this.props;
@@ -677,10 +682,11 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 			if (!Spotlight.focus(containerId)) {
 				Spotlight.setActiveContainer(containerId);
 			}
-		}
+		};
 
 		render () {
 			const {'data-webos-voice-exclusive': voiceExclusive, popupComponent: PopupComponent, popupClassName, noAutoDismiss, open, onClose, offset, popupProps, skin, spotlightRestrict, ...rest} = this.props;
+			const idFloatLayer = `${this.id}_floatLayer`;
 			let scrimType = rest.scrimType;
 			delete rest.scrimType;
 
@@ -711,8 +717,9 @@ const Decorator = hoc(defaultConfig, (config, Wrapped) => {
 			if (openProp) rest[openProp] = open;
 
 			return (
-				<div className={css.contextualPopupDecorator}>
+				<div aria-owns={idFloatLayer} className={css.contextualPopupDecorator}>
 					<FloatingLayer
+						id={idFloatLayer}
 						noAutoDismiss={noAutoDismiss}
 						onClose={this.handleClose}
 						onDismiss={onClose}