+++ internal/Picker/Picker.js
@@ -83,7 +83,7 @@ const allowedClassNames = ['picker', 'valueWrapper', 'joined', 'horizontal', 've
  */
 
 const PickerBase = class extends React.Component {
-	static displayName = 'Picker'
+	static displayName = 'Picker';
 
 	static propTypes = /** @lends sandstone/internal/Picker.Picker.prototype */ {
 		/**
@@ -401,7 +401,7 @@ const PickerBase = class extends React.Component {
 		 * @public
 		 */
 		wrap: PropTypes.bool
-	}
+	};
 
 	static defaultProps = {
 		accessibilityHint: '',
@@ -409,7 +409,7 @@ const PickerBase = class extends React.Component {
 		spotlightDisabled: false,
 		step: 1,
 		value: 0
-	}
+	};
 
 	constructor (props) {
 		super(props);
@@ -469,14 +469,14 @@ const PickerBase = class extends React.Component {
 		const shouldWrap = horizontalJoined || wrap;
 
 		return shouldWrap ? wrapRange(min, max, value + delta) : clamp(min, max, value + delta);
-	}
+	};
 
-	adjustDirection = (dir) => this.props.reverse ? -dir : dir
+	adjustDirection = (dir) => this.props.reverse ? -dir : dir;
 
 	hasReachedBound = (delta) => {
 		const {value} = this.props;
 		return this.computeNextValue(this.adjustDirection(delta)) === value;
-	}
+	};
 
 	updateValue = (dir) => {
 		const {disabled, onChange, step} = this.props;
@@ -486,7 +486,7 @@ const PickerBase = class extends React.Component {
 			const value = this.computeNextValue(dir * step);
 			onChange({value});
 		}
-	}
+	};
 
 	handleBlur = (ev) => {
 		forwardBlur(ev, this.props);
@@ -494,7 +494,7 @@ const PickerBase = class extends React.Component {
 		this.setState({
 			active: false
 		});
-	}
+	};
 
 	handleFocus = (ev) => {
 		forwardFocus(ev, this.props);
@@ -502,42 +502,42 @@ const PickerBase = class extends React.Component {
 		this.setState({
 			active: true
 		});
-	}
+	};
 
 	setTransitionDirection = (dir) => {
 		// change the transition direction based on the button press
 		this.reverseTransition = !(dir > 0);
-	}
+	};
 
 	handleDecrement = () => {
 		if (!this.hasReachedBound(-this.props.step)) {
 			this.updateValue(-1);
 			this.setPressedState(-1);
 		}
-	}
+	};
 
 	handleIncrement = () => {
 		if (!this.hasReachedBound(this.props.step)) {
 			this.updateValue(1);
 			this.setPressedState(1);
 		}
-	}
+	};
 
 	setPressedState = (pressed) => {
 		const {joined} = this.props;
 		if (joined) {
 			this.setState({pressed});
 		}
-	}
+	};
 
 	clearPressedState = () => {
 		this.pickerButtonPressed = 0;
 		this.setState({
 			pressed: 0
 		});
-	}
+	};
 
-	emulateMouseUp = new Job(this.clearPressedState, 175)
+	emulateMouseUp = new Job(this.clearPressedState, 175);
 
 	handleMouseDown = (ev) => {
 		const {joined, orientation} = this.props;
@@ -546,13 +546,13 @@ const PickerBase = class extends React.Component {
 		if (joined && orientation === 'horizontal') {
 			this.setIncPickerButtonPressed();
 		}
-	}
+	};
 
 	handleUp = () => {
 		if (this.props.joined && (this.pickerButtonPressed !== 0 || this.state.pressed !== 0)) {
 			this.emulateMouseUp.start();
 		}
-	}
+	};
 
 	handleDown = () => {
 		const {joined, orientation} = this.props;
@@ -567,12 +567,12 @@ const PickerBase = class extends React.Component {
 			this.handleDecrement();
 			this.emulateMouseUp.start();
 		}
-	}
+	};
 
 	handleIncDown = () => {
 		this.pickerButtonPressed = true;
 		this.handleIncrement();
-	}
+	};
 
 	handleWheel = (ev) => {
 		const {step} = this.props;
@@ -601,19 +601,19 @@ const PickerBase = class extends React.Component {
 				ev.stopPropagation();
 			}
 		}
-	}
+	};
 
-	throttleWheelInc = new Job(this.handleIncrement, 100)
+	throttleWheelInc = new Job(this.handleIncrement, 100);
 
-	throttleWheelDec = new Job(this.handleDecrement, 100)
+	throttleWheelDec = new Job(this.handleDecrement, 100);
 
 	setDecPickerButtonPressed = () => {
 		this.pickerButtonPressed = -1;
-	}
+	};
 
 	setIncPickerButtonPressed = () => {
 		this.pickerButtonPressed = 1;
-	}
+	};
 
 	handleHoldPulse = () => {
 		const {joined} = this.props;
@@ -622,7 +622,7 @@ const PickerBase = class extends React.Component {
 		} else if (joined && this.pickerButtonPressed === -1) {
 			this.handleDecrement();
 		}
-	}
+	};
 
 	handleKeyDown = (ev) => {
 		const {
@@ -661,7 +661,7 @@ const PickerBase = class extends React.Component {
 				onSpotlightRight(ev);
 			}
 		}
-	}
+	};
 
 	handleKeyUp = (ev) => {
 		const {
@@ -679,7 +679,7 @@ const PickerBase = class extends React.Component {
 				this.pickerButtonPressed = 0;
 			}
 		}
-	}
+	};
 
 	handleDecKeyDown = (ev) => {
 		const {keyCode} = ev;
@@ -705,7 +705,7 @@ const PickerBase = class extends React.Component {
 				forward(`onSpotlight${cap(direction)}`, ev, this.props);
 			}
 		}
-	}
+	};
 
 	handleIncKeyDown = (ev) => {
 		const {keyCode} = ev;
@@ -731,7 +731,7 @@ const PickerBase = class extends React.Component {
 				forward(`onSpotlight${cap(direction)}`, ev, this.props);
 			}
 		}
-	}
+	};
 
 	handleVoice = (ev) => {
 		const voiceIndex = ev && ev.detail && typeof ev.detail.matchedIndex !== 'undefined' && Number(ev.detail.matchedIndex);
@@ -744,7 +744,7 @@ const PickerBase = class extends React.Component {
 				ev.preventDefault();
 			}
 		}
-	}
+	};
 
 	determineClasses (css, decrementerDisabled, incrementerDisabled) {
 		const {className, joined, orientation, width} = this.props;