+++ samples/qa-a11y/src/views/ProgressBar.js
@@ -1,11 +1,10 @@
 import Button from '@enact/sandstone/Button';
-import Heading from '@enact/sandstone/Heading';
 import ProgressBar from '@enact/sandstone/ProgressBar';
 import React from 'react';
 
 import Section from '../components/Section';
 
-import css from '../App/App.module.less';
+import appCss from '../App/App.module.less';
 
 class ProgressBarView extends React.Component {
 	constructor () {
@@ -15,8 +14,8 @@ class ProgressBarView extends React.Component {
 		};
 	}
 
-	handleDecreaseBarValue = () => this.setState((state) => ({progressBarValue: Math.max((state.progressBarValue - 0.1).toFixed(1), 0)}))
-	handleIncreaseBarValue = () => this.setState((state) => ({progressBarValue: Math.min((state.progressBarValue + 0.1).toFixed(1), 1)}))
+	handleDecreaseBarValue = () => this.setState((state) => ({progressBarValue: Math.max((state.progressBarValue - 0.1).toFixed(1), 0)}));
+	handleIncreaseBarValue = () => this.setState((state) => ({progressBarValue: Math.min((state.progressBarValue + 0.1).toFixed(1), 1)}));
 
 	render = () => {
 		const {progressBarValue} = this.state;
@@ -29,21 +28,21 @@ class ProgressBarView extends React.Component {
 		}
 
 		return (
-			<>
-				<Heading showLine>Progress Bar</Heading>
-				<ProgressBar
-					aria-label={barAriaLabel}
-					aria-live="assertive"
-					progress={progressBarValue}
-				/>
-
-				<Section className={css.marginTop} title="Controls" vertical>
-					<Button alt="Decrease" aria-label="This is Decrease." icon="minus" onClick={this.handleDecreaseBarValue} />
-					<Button alt="Increase" aria-label="This is Increase." icon="plus" onClick={this.handleIncreaseBarValue} />
-				</Section>
-			</>
+			<Section title="Default">
+				<div>
+					<div className={appCss.controls}>
+						<Button aria-label="This is Decrease." icon="minus" onClick={this.handleDecreaseBarValue} />
+						<Button aria-label="This is Increase." icon="plus" onClick={this.handleIncreaseBarValue} />
+					</div>
+					<ProgressBar
+						aria-label={barAriaLabel}
+						aria-live="assertive"
+						progress={progressBarValue}
+					/>
+				</div>
+			</Section>
 		);
-	}
+	};
 }
 
 export default ProgressBarView;