+++ samples/qa-a11y/src/views/TooltipDecorator.js
@@ -7,7 +7,11 @@ import Section from '../components/Section';
 const TooltipButton = TooltipDecorator(Button);
 const tooltipProps = {'aria-hidden': true};
 
-const TooltipDecoratorView = () => (
+const TooltipDecoratorView = () => (<>
+	<Section title="Default">
+		<div alt="Normal">DO NOT USE the component wrapped with TooltipDecorator without `aria-label`.</div>
+	</Section>
+
 	<Section title="Aria-labelled">
 		<TooltipButton
 			alt="Below Right"
@@ -16,7 +20,7 @@ const TooltipDecoratorView = () => (
 			tooltipProps={tooltipProps}
 			tooltipText="Below Right Tooltip"
 		>
-			Below Right Tooltip
+			Text 0
 		</TooltipButton>
 
 		<TooltipButton
@@ -26,7 +30,7 @@ const TooltipDecoratorView = () => (
 			tooltipProps={tooltipProps}
 			tooltipText="Below Left Tooltip"
 		>
-			Below Left Tooltip
+			Text 1
 		</TooltipButton>
 
 		<TooltipButton
@@ -36,7 +40,7 @@ const TooltipDecoratorView = () => (
 			tooltipProps={tooltipProps}
 			tooltipText="Above Right Tooltip"
 		>
-			Above Right Tooltip
+			Text 2
 		</TooltipButton>
 
 		<TooltipButton
@@ -46,9 +50,9 @@ const TooltipDecoratorView = () => (
 			tooltipProps={tooltipProps}
 			tooltipText="Above Left Tooltip"
 		>
-			Above Left Tooltip
+			Text 3
 		</TooltipButton>
 	</Section>
-);
+</>);
 
 export default TooltipDecoratorView;