+++ samples/sampler/stories/qa/Dropdown.js
@@ -32,7 +32,7 @@ class AutoDismissDropdown extends React.Component {
 
 	handleClose = () => {
 		this.setState({open: false});
-	}
+	};
 
 	render () {
 		return (
@@ -59,7 +59,7 @@ class DisabledDropdown extends React.Component {
 
 	handleClick = () => {
 		this.setState({isDisabled: false});
-	}
+	};
 
 	render () {
 		return (