+++ samples/sampler/stories/qa/Scroller.js
@@ -51,7 +51,7 @@ class ScrollerResizableItem extends React.Component {
 		min: PropTypes.number,
 		more: PropTypes.bool,
 		toggleMore: PropTypes.func
-	}
+	};
 
 	render () {
 		const {max = 3000, min = 504, more, toggleMore} = this.props;
@@ -80,7 +80,7 @@ class ScrollerWithLongItem extends React.Component {
 
 	handleClick = () => {
 		this.setState(prevState => ({more: !prevState.more}));
-	}
+	};
 
 	render () {
 		return (
@@ -109,7 +109,7 @@ class ScrollerWithResizable extends React.Component {
 
 	handleClick = () => {
 		this.setState(prevState => ({more: !prevState.more}));
-	}
+	};
 
 	render () {
 		return (