+++ tests/ui/specs/Panel/PanelPage.js
@@ -1,5 +1,3 @@
-/* global document */
-
 'use strict';
 const {Page} = require('@enact/ui-test-utils/utils');
 
@@ -9,12 +7,22 @@ class PanelInterface {
 		this.id = `Panel${index}`;
 	}
 
-	get self () { return $(`#${this.id}`); }
-	get header () { return $(`#header${this.index}`); }
-	get body () { return $(`#body${this.index}`); }
+	get self () {
+		return $(`#${this.id}`);
+	}
+	get header () {
+		return $(`#header${this.index}`);
+	}
+	get body () {
+		return $(`#body${this.index}`);
+	}
 
-	get nextButton () { return $(`#next${this.index}`); }
-	get prevButton () { return $(`#prev${this.index}`); }
+	get nextButton () {
+		return $(`#next${this.index}`);
+	}
+	get prevButton () {
+		return $(`#prev${this.index}`);
+	}
 
 	waitForEnter () {
 		this.body.waitForExist();
@@ -29,6 +37,9 @@ class PanelPage extends Page {
 		this.panel2 = new PanelInterface(2);
 		this.panel3 = new PanelInterface(3);
 		this.panel4 = new PanelInterface(4);
+		this.panel5 = new PanelInterface(5);
+		this.panel6 = new PanelInterface(6);
+		this.panel7 = new PanelInterface(7);
 	}
 
 	focus (el) {