+++ tests/ui/specs/WizardPanels/WizardPanels-specs.js
@@ -40,18 +40,37 @@ describe('WizardPanels', function () {
 			wizardPanels.waitForLeave(2);
 			expect(wizardPanels.view1.isExisting()).to.be.true();
 		});
+
+		it('should focus on back button in header on left key', function () {
+			wizardPanels.focusNextButton();
+			Page.spotlightSelect();
+
+			wizardPanels.waitForLeave(1);
+			expect(wizardPanels.view2.isExisting()).to.be.true();
+
+			Page.spotlightUp();
+			expect(wizardPanels.prevButton.isFocused()).to.be.true();
+
+			Page.spotlightRight();
+			expect(wizardPanels.nextButton.isFocused()).to.be.true();
+
+			Page.spotlightLeft();
+			expect(wizardPanels.prevButton.isFocused()).to.be.true();
+
+		});
+
 	});
 
 	describe('Pointer', function () {
 		it('should navigate between views', function () {
-			browser.moveToElement(wizardPanels.nextButton.ELEMENT);
+			wizardPanels.nextButton.moveTo();
 			expect(wizardPanels.nextButton.isFocused()).to.be.true();
 			wizardPanels.nextButton.click();
 
 			wizardPanels.waitForLeave(1);
 			expect(wizardPanels.view2.isExisting()).to.be.true();
 
-			browser.moveToElement(wizardPanels.prevButton.ELEMENT);
+			wizardPanels.prevButton.moveTo();
 			expect(wizardPanels.prevButton.isFocused()).to.be.true();
 			wizardPanels.prevButton.click();